+++ activerecord/lib/active_record/connection_adapters/abstract_adapter.rb
@@ -739,7 +739,11 @@ module ActiveRecord
         def build_statement_pool
         end
 
-        def build_result(columns, rows, column_types = {})
+        # Builds the result object.
+        #
+        # This is an internal hook to make possible connection adapters to build
+        # custom result objects with connection-specific data.
+        def build_result(columns:, rows:, column_types: {})
           ActiveRecord::Result.new(columns, rows, column_types)
         end
     end